% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specgram.R
\name{audio2specgram}
\alias{audio2specgram}
\title{audio2specgram
Short Term Fourier Transform from time sequence}
\usage{
audio2specgram(x, config)
}
\arguments{
\item{x}{a vector of a time series}

\item{config}{configuration list generated by stftconfig().}
}
\value{
a matrix of time-frequency representation, where the row is the time and
    the column is the frequency. Number of the frequency bins are half of the window size
    plus one.
}
\description{
audio2specgram
Short Term Fourier Transform from time sequence
}
