% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stftconfig.R
\name{stftconfig}
\alias{stftconfig}
\title{stftconfig
Configuration of audio2specgram() and specgram2audio().}
\usage{
stftconfig(width, frameshift, towindow = hanning, fromwindow = rectwin)
}
\arguments{
\item{width}{number of sample points in an analysis window}

\item{frameshift}{number of sample points of frame shift}

\item{towindow}{an window function applied when calculating STFT.
The following functions can be used as 'towindow' and 'fromwindow' parameters:
rectwin: rectangle window
sinewin: sine window
hamming: hamming window (from signal package)
hanning: hanning window (from signal package)}

\item{fromwindow}{an window function applied when calculating inverse STFT}
}
\value{
a list of configuration parameters
}
\description{
stftconfig
Configuration of audio2specgram() and specgram2audio().
}
